refactor(opencode): extract SQLite compatibility parser
